Update Hale.Lib.ModuleLoader for xpat/dual-fw
Note: Compiles, but does not work correctly for NET Core
* Change target for Hale.Alert to ns2
