+++ Shared/Hale.Alert/Hale.Alert.csproj
@@ -1,6 +1,6 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFramework>net45</TargetFramework>
+    <TargetFramework>netstandard2</TargetFramework>
     <RootNamespace>Hale_Alert</RootNamespace>
     <AssemblyName>Hale-Alert</AssemblyName>
   </PropertyGroup>
