+++ Shared/Hale.Lib.ModuleLoader/ModuleLoader.cs
@@ -2,12 +2,11 @@
 {
     using System;
     using System.IO;
-    using System.Runtime.Remoting;
     using System.Security;
     using System.Security.Permissions;
     using Hale.Lib.Modules;
 
-    public class ModuleLoader : MarshalByRefObject
+    public partial class ModuleLoader : MarshalByRefObject
     {
         public static TResult ExecuteFunction<TResult>(string dll, string modulePath, string name, ModuleSettingsBase settings)
             where TResult : ModuleResultSet, new()
@@ -20,22 +19,6 @@
         {
             ModuleDomain moduleDomain = GetDomain(dll, modulePath);
             return moduleDomain.GetModuleInfo(Path.GetFullPath(Path.Combine(modulePath, dll)));
-        }
-
-        private static ModuleDomain GetDomain(string dll, string checkPath)
-        {
-            AppDomainSetup adSetup = new AppDomainSetup();
-
-            adSetup.ApplicationBase = Path.GetFullPath(checkPath);
-            PermissionSet permSet = new PermissionSet(PermissionState.Unrestricted);
-            AppDomain newDomain = AppDomain.CreateDomain("Module::" + Path.GetFileNameWithoutExtension(dll), null, adSetup, permSet);
-
-            ObjectHandle handle = Activator.CreateInstanceFrom(
-                newDomain,
-                typeof(ModuleDomain).Assembly.ManifestModule.FullyQualifiedName,
-                typeof(ModuleDomain).FullName);
-
-            return (ModuleDomain)handle.Unwrap();
         }
     }
 }
