+++ Shared/Hale.Lib.ModuleLoader/Partials.NetCore.cs
@@ -1,0 +1,14 @@
+﻿#if NETCOREAPP2_1
+
+namespace Hale.Lib.ModuleLoader
+{
+    public partial class ModuleLoader
+    {
+        private static ModuleDomain GetDomain(string dll, string checkPath)
+        {
+            return null;
+        }
+    }
+}
+
+#endif
